Updated the makeOverLapList() interface, and added a version that takes a BBox object.
Added the ComputeGeoBBox() method which attempts to guess if you
have a projected GeoTIFF or ISIS file, OR an unprojected ISIS file,
and does the right thing.  If it detects an unprojected ISIS file,
it calls the new ComputeGeoBBoxISIS().

Refactored ComputeGeoBoundary() to use ComputeGeoBBox() under the
hood, no change in interface.

Refactored makeOverlapList() to use the VW BBox object which makes
intersection calculations more straightforward.  There is still an
interface that takes a Vector4, so there is no change in the
interface.



git-svn-id: https://babelfish.arc.nasa.gov/svn/stereopipeline/sandbox/lima@1693 1395e39a-8119-0410-9602-abf82d98ffd9