In the process of writing a test for SaveStatistics() and ReadStatistics(), discovered a bug in SaveStatistics().
Cleanup on ReadStatistics(), interface unchanged.

The bug in SaveStatistics() resulted in any values that were greater
than the last element of histBins to potentially be counted 2x in
the last element of errorHist.  This was because of the looping
semantics: we looped on k, and then searched for stdv between
histBins[k] and histBins[k+1].  However, when k was on the last
element, k+1 was undefined.  On my system, it evaluted to 7.27186e+22,
and so numerically stdv was less than it, and it got wrongly added
to the bin.  It was easily fixed by starting k at 1, and then
evaluating on histBins[k-1] and histBins[k], which will always have
values, as long as histBins has at least two elements, which it
should.



git-svn-id: https://babelfish.arc.nasa.gov/svn/stereopipeline/sandbox/lima@1699 1395e39a-8119-0410-9602-abf82d98ffd9