Added a boost::program_options way of dealing with configuration settings.
This is just test code.  It is for discussion ONLY.  No actual
AlignTK programs should use this, because it would change the syntax
of the settings file that is currently in use.

Cons: 
(1) Would change the syntax of the settings file, equals signs are required  
          between name and value in the file.
(2) No multiple-valued options allowed, e.g.
       SAMPLING_STEP 8 8
       becomes:
       SAMPLING_STEP_H = 8
       SAMPLING_STEP_V = 8

This change in syntax comes from the requirements of using boost::program_options
to parse the file.

Pros:
(0) Provides more flexibility in setting options.
(1) Can override settings file parameters on the command line, e.g.
        ./alignTKprogram -d myDEM.tif -s example_settings.txt --SAMPLING_STEP_H 20
(2) Can also add command-line options to the settings file, e.g.
        in the settings file:
          MATCHING_MODE = 2
          inputDEMFiles = myDEM.tif
          SAMPLING_STEP_H = 8
          SAMPLING_STEP_V = 8
        then you can just run the above command like so, without the -d switch:
          ./alignTKprogram -s example_settings.txt
(3) We don't have to maintain text parsing code.
(4) Can change the defaults in the code, and have a new example settings file
        automatically written out.




git-svn-id: https://babelfish.arc.nasa.gov/svn/stereopipeline/sandbox/lima@1700 1395e39a-8119-0410-9602-abf82d98ffd9