Cleaned up the LOLA CSV file reading function, added a validity check.
Changed the guts of CSVFileRead() to LOLAFileRead(), since someday
we may also have a MOLAFileRead().  For the moment, CSVFileRead()
just calls LOLAFileRead(), but we should try to deprecate that.

Fixed some variable names in the pointCloud constructor argument that
were giving warnings on compile.

Aside from the addition of a bounding box validity check, not much was
changed in LOLAFileRead(), its working pretty well.



git-svn-id: https://babelfish.arc.nasa.gov/svn/stereopipeline/sandbox/lima@1718 1395e39a-8119-0410-9602-abf82d98ffd9