Cleaned up ComputeAllReflectance() and CHANGED its interface.
ComputeAllReflectance() had a CoregistrationParams struct as an
argument that was passed in, but it was never used within the
function.  I removed this as an argument.

This signature change (one less argument) required that all callers
of ComputeAllReflectance() be changed, too.  That is why lidar2img.cc
and tests/Test_tracks.cc are changed, but only in this surgical fashion.

The tests still pass.




git-svn-id: https://babelfish.arc.nasa.gov/svn/stereopipeline/sandbox/lima@1724 1395e39a-8119-0410-9602-abf82d98ffd9