Created a user tool to take a LOLA *PointPerRow_csv_table.csv file and break it into several individual CSV files that only contain one track per file.
I wanted to run lidar2dem on one track at a time.  Rather than get into the guts of lidar2dem and rebuild it, it was just simpler to create something that splits out all the tracks from a LOLA track file into individual files, and then you can just run lidar2dem (as is) on each of them individually.


git-svn-id: https://babelfish.arc.nasa.gov/svn/stereopipeline/sandbox/lima@1814 1395e39a-8119-0410-9602-abf82d98ffd9