We were getting Test errors because an edit to treat the origin as a special nodata value. That test has been removed.  Users of the ComputeMatchingError() should not pass bogus data into the function.
Initially, we thought about passing in a nodata for ComputeMatchingError()
to test against.  However, this requires an 'if' within that function,
and then in all the function calls downstream of that, since you must continually
test for a 'special' value, this resulted in poor performance.  Instead the check
was removed from ComputeMatchingError().

In order to properly deal with this situation, FindMatches() and
FindMatchesFromDEM() were adjusted to return a vector<bool> which
is a mask of 'good' values.  Then ICP_DEM_2_DEM() and ICP_DEM_2_LIDAR()
were adjusted to create temporary vectors built only from the 'good'
data.  These 'good' vectors were then used for ComputeMatchingError(),
find_centroid(), ComputeDEMRotation(), and ComputeDEMTranslation().



git-svn-id: https://babelfish.arc.nasa.gov/svn/stereopipeline/sandbox/lima@1984 1395e39a-8119-0410-9602-abf82d98ffd9